requirements.txt
argcomplete==2.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dbus-python==1.4.0
defusedxml==0.7.1
Flask==3.0.3
google-api-core==2.24.2
google-api-python-client==2.158.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==4.9.3
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pytesseract==0.3.13
requests==2.32.3
rsa==4.9
setuptools==78.1.0
soupsieve==2.6
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.2
wheel==0.45.1
